+++ api.tf
@@ -3,6 +3,12 @@ locals {
   // private gateway requires dualstack ip_address_type. for public gateways, ipv4 is the default. future improvement:
   // allow users to set this when deploying a public gateway
   gateway_ip_address_type = var.api_rest_type == "PRIVATE" ? "dualstack" : "ipv4"
+  # if var.custom_vpce_id was provided, the user is indicating they want to use a custom vpc endpoint which they have
+  # defined. if not, and if is_private_endpoint is true, we use the vpce we create in this module
+  vpce_id = var.custom_vpce_id != null ? var.custom_vpce_id : (local.is_private_endpoint ? aws_vpc_endpoint.stac_server_api_gateway_private[0].id : null)
+  # additionally, only create a vpc endpoint in this module if the api gateway is private *and* the user has not
+  # indicated they are using their own vpc endpoint
+  create_vpce = local.is_private_endpoint == true && var.custom_vpce_id == null
 }
 
 
@@ -90,7 +96,7 @@ resource "aws_vpc_security_group_ingress_rule" "stac_server_api_gateway_private_
 }
 
 resource "aws_vpc_endpoint" "stac_server_api_gateway_private" {
-  count = local.is_private_endpoint ? 1 : 0
+  count = local.create_vpce ? 1 : 0
 
   service_name        = "com.amazonaws.${data.aws_region.current.name}.execute-api"
   vpc_id              = var.vpc_id
@@ -114,7 +120,7 @@ resource "aws_api_gateway_rest_api" "stac_server_api_gateway" {
 
   endpoint_configuration {
     types            = [var.api_rest_type]
-    vpc_endpoint_ids = local.is_private_endpoint ? aws_vpc_endpoint.stac_server_api_gateway_private[*].id : null
+    vpc_endpoint_ids = local.is_private_endpoint ? [local.vpce_id] : null
     ip_address_type  = local.gateway_ip_address_type
   }
 }
@@ -136,7 +142,7 @@ data "aws_iam_policy_document" "stac_server_api_gateway_private" {
     condition {
       variable = "aws:SourceVpce"
       test     = "StringNotEquals"
-      values   = [aws_vpc_endpoint.stac_server_api_gateway_private[0].id]
+      values   = [local.vpce_id]
     }
   }
 
@@ -362,7 +368,7 @@ resource "aws_api_gateway_domain_name" "stac_server_api_gateway_domain_name" {
       "Resource": "arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:/domainnames/*",
       "Condition": {
         "StringNotEquals": {
-          "aws:SourceVpce": "${aws_vpc_endpoint.stac_server_api_gateway_private[0].id}"
+          "aws:SourceVpce": "${local.vpce_id}"
         }
       }
     }
@@ -373,7 +379,7 @@ EOF
 
 resource "aws_api_gateway_domain_name_access_association" "titiler_api_gateway_domain_name_access_association" {
   count                          = local.is_private_endpoint == true && var.domain_alias != "" && var.private_certificate_arn != "" ? 1 : 0
-  access_association_source      = aws_vpc_endpoint.stac_server_api_gateway_private[0].id
+  access_association_source      = local.vpce_id
   access_association_source_type = "VPCE"
   domain_name_arn                = aws_api_gateway_domain_name.stac_server_api_gateway_domain_name[0].arn
 }